#8 Refactor sample front end to enable testing with pytest-qt
